+++ include/seqan3/alignment/scoring/all.hpp
@@ -19,7 +19,6 @@
  */
 
 #include <seqan3/alignment/scoring/aminoacid_scoring_scheme.hpp>
-#include <seqan3/alignment/scoring/gap_scheme.hpp>
 #include <seqan3/alignment/scoring/nucleotide_scoring_scheme.hpp>
 #include <seqan3/alignment/scoring/scoring_scheme_base.hpp>
 #include <seqan3/alignment/scoring/scoring_scheme_concept.hpp>